requirements.txt
django==5.1.4
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2==2.9.10
django-cors-headers==4.6.0
drf-yasg==1.21.8
djangorestframework-simplejwt==5.4.0
gunicorn
whitenoise
dj-database-url
